I think I got it with the FortranMatrix thing

+++ SHARP3D/SHARP3D/Utils/Enum/DataLength.cs
@@ -5,6 +5,6 @@
         CHAR = -1,
         BYTE = 1,
         INT16 = 2,
-        INT32 = 4,
+        FLOAT32 = 4,
     }
 }
